simcore multiply fix. GUI frames reorganization

+++ reference/plp-3.0/sw/PLPTool/src/plptool/mips/SimCore.java
@@ -1227,8 +1227,8 @@ public class SimCore extends PLPSimCore {
                             return (a < b) ? 1 : 0;
                         case 0x00: return b << MIPSInstr.sa(instr);
                         case 0x02: return b >>> MIPSInstr.sa(instr);
-			case 0x10: return (a * b) & 0xffffffffL;
-			case 0x11: return (((int) a * (int) b) & 0xffffffff00000000L) >> 32;
+			case 0x10: return ((long)(int) a * (long)(int)b) & 0xffffffffL;
+			case 0x11: return (((long)(int) a * (long)(int) b) & 0xffffffff00000000L) >> 32;
                     }
 
                 case 0x04: return (a - b == 0) ? 1 : 0;
